Add A Starter Home Page & Components Folder
I've created a components folder for storing most
of our react components. In that folder, I created
a `Home` folder. This is for storing code related to
what we might refer to as "the home page." I've put
a basic starter component (that simply says it will
be the home page component) in Home.js within the
`Home` folder. I've also put an empty Home.css file.

Also within the components folder, I added a file
called `components.js`. This file is for managing
exporting components out of the components folder.
Doing this means that we can use a single import
statement for importing multiple components. In
otherwords, we won't need an import statement
for each individual component. We'll just need
to import each individual component from one file.