Add Routing Logic
Wrap the app in a react router so that we can switch what
component we display depending on the url. Add a route for
the previously added (though not completed) `Home` component.
Now, when the path is exactly `/`, the `Home` component displays.
Any other path doesn't display anything below the space reserved
for the header.