Create Frame Around Page Content
Create a frame that will go around all the content we will be showing
users. In other words, this frame goes around the, "main content," and
we'll have our main navigation to the left of this.

Note: we'll  want to end up changing the `height` of the frame to
"fit-content" later on, and we'll want to make its `width` dependent
upon how much space the side bar ends up taking.