Make Frame Size Responsive
Make the frame size responsive, because our side bar is going to change
sizes depending on if its expanded or minimized. We'll want the frame to
adjust accordingly. Also, leave a comment to indicate where the side bar
component should get inserted in order for it to work inside the flexbox
container.