Creating Initial Scaffolding for Style
Set up an outline to start creating the `FileList` component. This
instantiates the size, shape, and rough look of the component overall.
Place it on the vendor app. Eventually, we'll want to use tabs to
determine which feature to show when. However, since tabs haven't yet
been added here, just use a bool to only display what I want to see for
the sake of development.