Add Scrolling to the File List
In the event that the file list is very long, we'll want a way for the
user to scroll through the files without necessarily scrolling through
the entire page. Add a scroll bar when the list goes past the page
height. Also, edit the scroll bar so that it somewhat matches the style
we've been going for.

Note: this sort of style change using webkit isn't supported on ie11, so
ie users will see a normal scrollbar.